fix: refactor and bug fixes
- Moved login to auth
- Add test for duplicate account
